+++ src/EntityGraphQL/Schema/FieldExtensions/Sorting/UseSortExtension.cs
@@ -18,7 +18,40 @@ namespace EntityGraphQL.Schema.FieldExtensions
         {
             if (!field.Resolve.Type.IsEnumerableOrArray())
                 throw new ArgumentException($"UseSort must only be called on a field that returns an IEnumerable");
-            field.AddExtension(new SortExtension(fieldSelection.ReturnType));
+            field.AddExtension(new SortExtension(fieldSelection.ReturnType, null, null));
+            return field;
+        }
+        /// <summary>
+        /// Update field to implement a sort argument that takes options to sort a collection
+        /// Only call on a field that returns an IEnumerable
+        /// </summary>
+        /// <param name="field"></param>
+        /// <param name="fieldSelection">Select the fields you want available for sorting.
+        /// T must be the context of the collection you are applying the sort to</param>
+        /// <param name="defaultSort">Sort to use if no sort argument supplied in query</param>
+        /// <param name="direction">Direction of the default sort</param>
+        /// <returns></returns>
+        public static Field UseSort<ElementType, ReturnType, TSort>(this Field field, Expression<Func<ElementType, ReturnType>> fieldSelection, Expression<Func<ElementType, TSort>> defaultSort, SortDirectionEnum direction)
+        {
+            if (!field.Resolve.Type.IsEnumerableOrArray())
+                throw new ArgumentException($"UseSort must only be called on a field that returns an IEnumerable");
+            field.AddExtension(new SortExtension(fieldSelection.ReturnType, defaultSort, direction));
+            return field;
+        }
+
+        /// <summary>
+        /// Update field to implement a sort argument that takes options to sort a collection
+        /// Only call on a field that returns an IEnumerable
+        /// </summary>
+        /// <param name="field"></param>
+        /// <param name="defaultSort">Sort to use if no sort argument supplied in query</param>
+        /// <param name="direction">Direction of the default sort</param>
+        /// <returns></returns>
+        public static Field UseSort<ElementType, TSort>(this Field field, Expression<Func<ElementType, TSort>> defaultSort, SortDirectionEnum direction)
+        {
+            if (!field.Resolve.Type.IsEnumerableOrArray())
+                throw new ArgumentException($"UseSort must only be called on a field that returns an IEnumerable");
+            field.AddExtension(new SortExtension(null, defaultSort, direction));
             return field;
         }
 
@@ -32,7 +65,7 @@ namespace EntityGraphQL.Schema.FieldExtensions
         {
             if (!field.Resolve.Type.IsEnumerableOrArray())
                 throw new ArgumentException($"UseSort must only be called on a field that returns an IEnumerable");
-            field.AddExtension(new SortExtension(null));
+            field.AddExtension(new SortExtension(null, null, null));
             return field;
         }
     }
@@ -40,11 +73,6 @@ namespace EntityGraphQL.Schema.FieldExtensions
     public class UseSortAttribute : FieldExtensionAttribute
     {
         public UseSortAttribute() { }
-
-        public UseSortAttribute(Func<Expression> fields)
-        {
-
-        }
 
         public override void ApplyExtension(Field field)
         {